refactor: get rid of .sync and checkoutv2
