+++ packages/commercetools/theme/components/Checkout/UserBillingAddresses.vue
@@ -29,7 +29,7 @@
     <SfCheckbox
       data-cy="billing-details-checkbox_isDefault"
       :selected="setAsDefault"
-      @change="$emit('update:setAsDefault', $event)"
+      @change="$emit('input', $event)"
       name="setAsDefault"
       label="Use this address as my default one."
       class="billing-address-setAsDefault"
