+++ packages/commercetools/theme/pages/Checkout/Payment.vue
@@ -1,31 +1,42 @@
 <template>
-  <ValidationObserver v-slot="{ handleSubmit }">
-    <form @submit.prevent="handleSubmit(addressIsModified ? handleBillingAddressSubmit : handleFormSubmit)">
+  <ValidationObserver v-slot="{ handleSubmit, dirty, reset }">
     <SfHeading
       :level="3"
-      title="Billing address"
+      :title="$t('Billing')"
       class="sf-heading--left sf-heading--no-underline title"
     />
-    <UserBillingAddresses
-      v-if="isAuthenticated && billingAddresses && billingAddresses.length"
-      :setAsDefault="setAsDefault"
-      :billingAddresses="billingAddresses"
-      :currentAddressId="currentAddressId"
-      @setCurrentAddress="setCurrentAddress($event)"
-      @changeSetAsDefault="setAsDefault = $event"
-    />
-    <SfCheckbox
-      :selected="sameAsShipping"
-      @change="handleCheckSameAddress"
-      label="Copy address data from shipping"
-      name="copyShippingAddress"
-      class="form__element"
-    />
+    <form
+      @submit.prevent="
+        handleSubmit(
+          isBillingDetailsCompleted && !dirty
+            ? handleStepSubmit
+            : handleAddressSubmit(reset)
+        )
+      "
+    >
+      <UserBillingAddresses
+        v-if="isAuthenticated && hasSavedBillingAddress"
+        v-model="setAsDefault"
+        :currentAddressId="currentAddressId || NOT_SELECTED_ADDRESS"
+        @setCurrentAddress="handleSetCurrentAddress"
+      />
+      <SfCheckbox
+        :selected="sameAsShipping"
+        @change="handleCheckSameAddress"
+        label="Copy address data from shipping"
+        name="copyShippingAddress"
+        class="form__element"
+      />
       <div class="form" v-if="canAddNewAddress">
-        <ValidationProvider name="firstName" rules="required|min:2" v-slot="{ errors }" slim>
+        <ValidationProvider
+          name="firstName"
+          rules="required|min:2"
+          v-slot="{ errors }"
+          slim
+        >
           <SfInput
             :value="billingDetails.firstName"
-            @input="firstName => setBillingDetailsAndUnpickAddress({ firstName })"
+            @input="firstName => changeDetails('firstName', firstName)"
             label="First name"
             name="firstName"
             class="form__element form__element--half"
@@ -34,10 +45,15 @@
             :errorMessage="errors[0]"
           />
         </ValidationProvider>
-        <ValidationProvider name="lastName" rules="required|min:2" v-slot="{ errors }" slim>
+        <ValidationProvider
+          name="lastName"
+          rules="required|min:2"
+          v-slot="{ errors }"
+          slim
+        >
           <SfInput
             :value="billingDetails.lastName"
-            @input="lastName => setBillingDetailsAndUnpickAddress({ lastName })"
+            @input="lastName => changeDetails('lastName', lastName)"
             label="Last name"
             name="lastName"
             class="form__element form__element--half form__element--half-even"
@@ -46,10 +62,15 @@
             :errorMessage="errors[0]"
           />
         </ValidationProvider>
-        <ValidationProvider name="streetName" rules="required|min:2" v-slot="{ errors }" slim>
+        <ValidationProvider
+          name="streetName"
+          rules="required|min:2"
+          v-slot="{ errors }"
+          slim
+        >
           <SfInput
             :value="billingDetails.streetName"
-            @input="streetName => setBillingDetailsAndUnpickAddress({ streetName })"
+            @input="streetName => changeDetails('streetName', streetName)"
             label="Street name"
             name="streetName"
             class="form__element form__element--half"
@@ -58,10 +79,15 @@
             :errorMessage="errors[0]"
           />
         </ValidationProvider>
-        <ValidationProvider name="apartment" rules="required|min:2" v-slot="{ errors }" slim>
+        <ValidationProvider
+          name="apartment"
+          rules="required|min:2"
+          v-slot="{ errors }"
+          slim
+        >
           <SfInput
-            :value="billingDetails.streetNumber"
-            @input="streetNumber => setBillingDetailsAndUnpickAddress({ streetNumber })"
+            :value="billingDetails.apartment"
+            @input="apartment => changeDetails('apartment', apartment)"
             label="House/Apartment number"
             name="apartment"
             class="form__element form__element--half form__element--half-even"
@@ -70,10 +96,15 @@
             :errorMessage="errors[0]"
           />
         </ValidationProvider>
-        <ValidationProvider name="city" rules="required|min:2" v-slot="{ errors }" slim>
+        <ValidationProvider
+          name="city"
+          rules="required|min:2"
+          v-slot="{ errors }"
+          slim
+        >
           <SfInput
             :value="billingDetails.city"
-            @input="city => setBillingDetailsAndUnpickAddress({ city })"
+            @input="city => changeDetails('city', city)"
             label="City"
             name="city"
             class="form__element form__element--half"
@@ -82,10 +113,52 @@
             :errorMessage="errors[0]"
           />
         </ValidationProvider>
-        <ValidationProvider name="zipCode" rules="required|min:2" v-slot="{ errors }" slim>
+        <ValidationProvider
+          name="state"
+          slim
+        >
+          <SfInput
+            :value="billingDetails.state"
+            @input="state => changeDetails('state', state)"
+            label="State/Province"
+            name="state"
+            class="form__element form__element--half form__element--half-even"
+          />
+        </ValidationProvider>
+        <ValidationProvider
+          name="country"
+          rules="required|min:2"
+          v-slot="{ errors }"
+          slim
+        >
+          <SfSelect
+            :value="billingDetails.country"
+            @input="country => changeDetails('country', country)"
+            label="Country"
+            name="country"
+            class="form__element form__element--half form__select sf-select--underlined"
+            required
+            :valid="!errors[0]"
+            :errorMessage="errors[0]"
+          >
+            <SfSelectOption
+              v-for="countryOption in countries"
+              :key="countryOption.name"
+              :value="countryOption.name"
+            >
+              {{ countryOption.label }}
+            </SfSelectOption>
+          </SfSelect>
+        </ValidationProvider>
+        <ValidationProvider
+          name="zipCode"
+          rules="required|min:2"
+          v-slot="{ errors }"
+          slim
+        >
           <SfInput
             :value="billingDetails.postalCode"
-            @input="postalCode => setBillingDetailsAndUnpickAddress({ postalCode })"
+            @input="postalCode => changeDetails('postalCode', postalCode)"
             label="Zip-code"
             name="zipCode"
             class="form__element form__element--half form__element--half-even"
@@ -94,32 +167,18 @@
             :errorMessage="errors[0]"
           />
         </ValidationProvider>
-        <ValidationProvider name="country" rules="required|min:2" v-slot="{ errors }" slim>
-          <SfSelect
-            :value="billingDetails.country"
-            @input="country => setBillingDetailsAndUnpickAddress({ country })"
-            label="Country"
-            class="form__element form__element--half form__select sf-select--underlined"
-            required
-            :valid="!errors[0]"
-            :errorMessage="errors[0]"
-          >
-            <SfSelectOption
-              v-for="countryOption in COUNTRIES"
-              :key="countryOption.key"
-              :value="countryOption.key"
-            >
-              {{ countryOption.label }}
-            </SfSelectOption>
-          </SfSelect>
-        </ValidationProvider>
-        <ValidationProvider name="phone" rules="required|min:2" v-slot="{ errors }" slim>
+        <ValidationProvider
+          name="phone"
+          rules="required|digits:9"
+          v-slot="{ errors }"
+          slim
+        >
           <SfInput
             :value="billingDetails.phone"
-            @input="phone => setBillingDetailsAndUnpickAddress({ contactInfo: { phone } })"
+            @input="phone => changeDetails('phone', phone)"
             label="Phone number"
             name="phone"
-            class="form__element form__element--half form__element--half-even"
+            class="form__element form__element--half"
             required
             :valid="!errors[0]"
             :errorMessage="errors[0]"
@@ -128,50 +187,41 @@
       </div>
       <SfButton
         v-if="!canAddNewAddress"
-        class="form__action-button form__action-button--margin-bottom"
+        class="color-light form__action-button form__action-button--add-address"
         type="submit"
-        @click.native="canAddNewAddress = true"
+        @click.native="handleAddNewAddressBtnClick"
       >
         {{ $t('Add new address') }}
       </SfButton>
-    <SfHeading
-      v-if="canContinueToReview"
-      :level="3"
-      title="Payment methods"
-      class="sf-heading--left sf-heading--no-underline title"
-    />
-    <div class="form">
-      <div class="form__element payment-methods" v-if="canContinueToReview">
-        <SfRadio
-          v-for="item in paymentMethods"
-          :key="item.value"
-          :selected="chosenPaymentMethod.value"
-          :label="item.label"
-          :value="item.value"
-          @input="setPaymentMethod(item, { save: true })"
-          name="paymentMethod"
-          :description="item.description"
-          class="form__radio payment-method"
-        >
-          <template #label>
-            <div class="sf-radio__label">
-              {{ item.label }}
-            </div>
-          </template>
-        </SfRadio>
+      <SfHeading
+        v-if="isBillingDetailsCompleted && !dirty"
+        :level="3"
+        title="Payment methods"
+        class="sf-heading--left sf-heading--no-underline title"
+      />
+      <div class="form">
+        <div class="form__action">
+          <nuxt-link
+            to="/checkout/personal-details"
+            class="sf-button color-secondary form__back-button"
+            >Go back</nuxt-link>
+          <SfButton
+            class="form__action-button"
+            type="submit"
+            v-if="isBillingDetailsCompleted && !dirty"
+          >
+            {{ $t('Continue to payment') }}
+          </SfButton>
+          <SfButton
+            class="form__action-button"
+            type="submit"
+            :disabled="loading"
+            v-else
+          >
+            {{ $t('Select payment method') }}
+          </SfButton>
+        </div>
       </div>
-      <div class="form__action">
-        <nuxt-link to="/checkout/shipping" class="sf-button color-secondary form__back-button">
-          {{ $t('Go back') }}
-        </nuxt-link>
-        <SfButton class="form__action-button" type="submit" :disabled="loading.billingAddress" v-if="canContinueToReview">
-          {{ $t('Review my order') }}
-        </SfButton>
-        <SfButton class="form__action-button" type="submit" :disabled="loading.billingAddress" v-else>
-          {{ $t('Select payment method') }}
-        </SfButton>
-      </div>
-    </div>
     </form>
   </ValidationObserver>
 </template>
@@ -183,25 +233,17 @@ import {
   SfButton,
   SfSelect,
   SfRadio,
-  SfImage,
   SfCheckbox
 } from '@storefront-ui/vue';
-import { ref, computed, onMounted } from '@vue/composition-api';
-import { useCheckout, useUser, userBillingGetters, useUserBilling } from '@vue-storefront/commercetools';
+import { useUserBilling, userBillingGetters, useUser, useBilling, useShipping } from '@vue-storefront/commercetools';
 import { ValidationProvider, ValidationObserver, extend } from 'vee-validate';
-import { required, min } from 'vee-validate/dist/rules';
+import { required, min, digits } from 'vee-validate/dist/rules';
+import { useVSFContext } from '@vue-storefront/core';
+import { ref, watch, computed, onMounted } from '@vue/composition-api';
 import { onSSR } from '@vue-storefront/core';
 
-const COUNTRIES = [
-  { key: 'US',
-    label: 'United States' },
-  { key: 'UK',
-    label: 'United Kingdom' },
-  { key: 'IT',
-    label: 'Italy' },
-  { key: 'PL',
-    label: 'Poland' }
-];
+const NOT_SELECTED_ADDRESS = '';
+
 extend('required', {
   ...required,
   message: 'This field is required'
@@ -210,154 +252,162 @@ extend('min', {
   ...min,
   message: 'The field should have at least {length} characters'
 });
+extend('digits', {
+  ...digits,
+  message: 'Please provide a valid phone number'
+});
+
 export default {
-  name: 'Payment',
+  name: 'Billing',
   components: {
     SfHeading,
     SfInput,
     SfButton,
     SfSelect,
     SfRadio,
-    SfImage,
     SfCheckbox,
     UserBillingAddresses: () => import('~/components/Checkout/UserBillingAddresses'),
     ValidationProvider,
     ValidationObserver
   },
-  setup(props, context) {
-    const {
-      billingDetails,
-      shippingDetails,
-      paymentMethods,
-      chosenPaymentMethod,
-      loadPaymentMethods,
-      isBillingAddressCompleted,
-      setBillingDetails,
-      setPaymentMethod,
-      loadDetails,
-      loading
-    } = useCheckout();
-    const { billing, load: loadUserBilling, setDefaultAddress } = useUserBilling();
+  props: {
+    handleBillingAddressSubmit: Function
+  },
+  setup(_, context) {
+    const { $ct: { config } } = useVSFContext();
+    const { shipping: shippingDetails, load: loadShipping } = useShipping();
+    const { billing: address, loading, load, save } = useBilling();
     const { isAuthenticated } = useUser();
+    const { billing: userBilling, load: loadUserBilling, setDefaultAddress } = useUserBilling();
+    const billingDetails = ref(address.value || {});
 
-    const canAddNewAddress = ref(true);
-    const addressIsModified = ref(false);
-    const currentAddressId = ref(-1);
+    const isBillingMethodCompleted = ref(false);
+    const isBillingDetailsCompleted = ref(false);
+    const currentAddressId = ref(NOT_SELECTED_ADDRESS);
     const setAsDefault = ref(false);
-
+    const canAddNewAddress = ref(true);
     const sameAsShipping = ref(false);
     let oldBilling = null;
 
-    const mapAbstractAddressToIntegrationAddress = address => ({
-      ...billingDetails.value,
-      contactInfo: {
-        ...billingDetails.value.contactInfo,
-        phone: userBillingGetters.getPhone(address)
-      },
-      streetNumber: userBillingGetters.getApartmentNumber(address),
-      city: userBillingGetters.getCity(address),
-      country: userBillingGetters.getCountry(address),
-      state: userBillingGetters.getProvince(address),
-      firstName: userBillingGetters.getFirstName(address),
-      lastName: userBillingGetters.getLastName(address),
-      streetName: userBillingGetters.getStreetName(address),
-      postalCode: userBillingGetters.getPostCode(address)
+    const hasSavedBillingAddress = computed(() => {
+      if (!isAuthenticated.value || !userBilling.value) {
+        return false;
+      }
+      const addresses = userBillingGetters.getAddresses(userBilling.value);
+      return Boolean(addresses?.length);
     });
 
-    const setCurrentAddress = async (addressId) => {
-      const chosenAddress = userBillingGetters.getAddresses(billing.value, { id: addressId });
-      if (!chosenAddress || !chosenAddress.length) {
+    const handleCheckSameAddress = async () => {
+      sameAsShipping.value = !sameAsShipping.value;
+      if (sameAsShipping.value) {
+        if (!shippingDetails.value) {
+          await loadShipping();
+        }
+        oldBilling = {...billingDetails.value};
+        billingDetails.value = {...shippingDetails.value};
+        currentAddressId.value = NOT_SELECTED_ADDRESS;
+        setAsDefault.value = false;
+        isBillingDetailsCompleted.value = false;
         return;
       }
-      currentAddressId.value = addressId;
-      setBillingDetails(mapAbstractAddressToIntegrationAddress(chosenAddress[0]));
-      addressIsModified.value = true;
+      isBillingDetailsCompleted.value = false;
+      billingDetails.value = oldBilling;
+    };
+
+    const handleStepSubmit = () => context.emit('submit');
+
+    const handleAddressSubmit = (reset) => async () => {
+      const addressId = currentAddressId.value;
+      await save({ billingDetails: billingDetails.value });
+      if (addressId !== NOT_SELECTED_ADDRESS && setAsDefault.value) {
+        const chosenAddress = userBillingGetters.getAddresses(userBilling.value, { id: addressId });
+        if (chosenAddress && chosenAddress.length) {
+          await setDefaultAddress({ address: chosenAddress[0] });
+        }
+      }
+      reset();
+      isBillingDetailsCompleted.value = true;
+    };
+
+    const handleAddNewAddressBtnClick = () => {
+      currentAddressId.value = NOT_SELECTED_ADDRESS;
+      canAddNewAddress.value = true;
+    };
+
+    const handleSetCurrentAddress = address => {
+      billingDetails.value = {...address};
+      currentAddressId.value = address.id;
+      canAddNewAddress.value = false;
+      isBillingDetailsCompleted.value = false;
+      isBillingMethodCompleted.value = false;
       sameAsShipping.value = false;
     };
 
+    const changeDetails = (field, value) => {
+      billingDetails.value[field] = value;
+      isBillingMethodCompleted.value = false;
+      currentAddressId.value = NOT_SELECTED_ADDRESS;
+    };
+
+    const selectDefaultAddress = () => {
+      const defaultAddress = userBillingGetters.getAddresses(userBilling.value, { isDefault: true });
+      if (defaultAddress && defaultAddress.length) {
+        handleSetCurrentAddress(defaultAddress[0]);
+      }
+    };
+
+    // Update local state if we have new address' response from the backend
+    watch(address, (addr) => {
+      billingDetails.value = addr;
+    });
+
     onSSR(async () => {
-      await loadDetails();
-      await loadPaymentMethods();
+      await load();
+      if (isAuthenticated.value) {
+        await loadUserBilling();
+      }
     });
 
     onMounted(async () => {
-      if (isAuthenticated.value) {
+      if (!userBilling.value?.addresses && isAuthenticated.value) {
         await loadUserBilling();
-        const billingAddresses = userBillingGetters.getAddresses(billing.value);
-        if (!billingAddresses || !billingAddresses.length) {
-          return;
-        }
-        canAddNewAddress.value = false;
-        if (userBillingGetters.isDefault(billingAddresses[0])) {
-          setCurrentAddress(userBillingGetters.getId(billingAddresses[0]));
-        }
       }
-    });
-
-    const handleFormSubmit = async () => {
-      await setBillingDetails(billingDetails.value, { save: true });
-      context.root.$router.push('/checkout/order-review');
-    };
-
-    const handleBillingAddressSubmit = async () => {
-      await setBillingDetails(billingDetails.value, { save: true });
-      // here we should have something like loadPaymentMethods which depends on billing address imo
-      if (currentAddressId.value > -1 && setAsDefault.value) {
-        const chosenAddress = userBillingGetters.getAddresses(billing.value, { id: currentAddressId.value });
-        if (!chosenAddress || !chosenAddress.length) {
-          return;
-        }
-        await setDefaultAddress(chosenAddress[0]);
-      }
-      addressIsModified.value = false;
-    };
-
-    const handleCheckSameAddress = () => {
-      sameAsShipping.value = !sameAsShipping.value;
-      if (sameAsShipping.value) {
-        oldBilling = billingDetails.value;
-        setBillingDetails(shippingDetails.value);
-        currentAddressId.value = -1;
-        setAsDefault.value = false;
-        addressIsModified.value = true;
+      const billingAddresses = userBillingGetters.getAddresses(userBilling.value);
+      if (!billingAddresses || !billingAddresses.length) {
         return;
       }
-      setBillingDetails(oldBilling);
-    };
-
-    const setBillingDetailsAndUnpickAddress = value => {
-      setBillingDetails(value);
-      currentAddressId.value = -1;
-      addressIsModified.value = true;
-      sameAsShipping.value = false;
-    };
+      const hasEmptyBillingDetails = !billingDetails.value || Object.keys(billingDetails.value).length === 0;
+      if (hasEmptyBillingDetails) {
+        selectDefaultAddress();
+        return;
+      }
+      canAddNewAddress.value = false;
+    });
 
     return {
-      loading,
-      billingDetails,
-      paymentMethods,
-      chosenPaymentMethod,
-      sameAsShipping,
-      COUNTRIES,
-      setPaymentMethod,
-      handleFormSubmit,
-      handleBillingAddressSubmit,
-      handleCheckSameAddress,
+      NOT_SELECTED_ADDRESS,
       isAuthenticated,
-      billingAddresses: computed(() => userBillingGetters.getAddresses(billing.value)),
+      billingDetails,
+      countries: config.countries,
       setAsDefault,
-      canContinueToReview: computed(() => isBillingAddressCompleted.value && !addressIsModified.value),
-      currentAddressId: computed(() => currentAddressId.value),
-      setCurrentAddress,
-      addressIsModified,
-      setBillingDetailsAndUnpickAddress,
       canAddNewAddress,
-      userBillingGetters
+      currentAddressId,
+      hasSavedBillingAddress,
+      isBillingMethodCompleted,
+      isBillingDetailsCompleted,
+      handleAddressSubmit,
+      handleStepSubmit,
+      handleAddNewAddressBtnClick,
+      handleSetCurrentAddress,
+      handleCheckSameAddress,
+      changeDetails,
+      sameAsShipping,
+      shippingDetails,
+      loading
     };
   }
 };
 </script>
-
 <style lang="scss" scoped>
 .title {
   margin: var(--spacer-xl) 0 var(--spacer-base) 0;
@@ -418,6 +468,14 @@ export default {
         order: -1;
         --button-margin: 0;
         text-align: left;
+      }
+    }
+     &--add-address {
+      width: 100%;
+      margin: 0;
+      @include for-desktop {
+        margin: 0 0 var(--spacer-lg) 0;
+        width: auto;
       }
     }
   }
